+++ CHANGES.md
@@ -7,6 +7,10 @@
   separate repositories.
 - Library installation has been cleaned up a bit.
 - List indexing now gives an optional value
+- Added a `--format` flag to the `tomo` binary that autoformats your code
+  (currently unstable, do not rely on it just yet).
+- Fixed bugs:
+  - `Int.parse()` had a memory bug.
 
 ## v0.3